Fix DB configuration to use asyncronous and fix imports with isort
Database now using asyncronous configuration, endpoints tested with it.
Commit does not include length check, will be coming but not yet too
relevant. Ran `black`, `flake8` and `isort` against the code to ensure
that it would follow pep8. This message does not probably include
everything that was changed, but it is a good start.